backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]